Add ExtIOFunc struct to manage external IO
