+++ cmd_test.go
@@ -3,14 +3,11 @@ package main_test
 import (
 	"bufio"
 	"bytes"
-	"fmt"
 	"io"
 	"os"
-	"reflect"
 	"strings"
 	"testing"
 
-	"github.com/keybase/go-keychain"
 	. "github.com/m-mizutani/altenv"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -30,20 +27,22 @@ func toEnvVars(buf *bytes.Buffer) map[string]string { // nolint
 
 func makeParameters(buf *bytes.Buffer) *Parameters {
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "my.env":
-				return ToReadCloser("COLOR=BLUE"), nil
-			case "my2.env":
-				return ToReadCloser("COSMOS=STARS"), nil
-			case "my.json":
-				return ToReadCloser(`{"MAGIC":"5"}`), nil
-			case "my2.json":
-				return ToReadCloser(`{"COLOR":"orange"}`), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "my.env":
+					return ToReadCloser("COLOR=BLUE"), nil
+				case "my2.env":
+					return ToReadCloser("COSMOS=STARS"), nil
+				case "my.json":
+					return ToReadCloser(`{"MAGIC":"5"}`), nil
+				case "my2.json":
+					return ToReadCloser(`{"COLOR":"orange"}`), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 
@@ -171,22 +170,24 @@ func newConfigTestApp(buf *bytes.Buffer) *cli.App {
 `
 
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "testconfig":
-				return ToReadCloser(configData), nil
-			case "envfile_global_1.env":
-				return ToReadCloser("COLOR=BLUE"), nil
-			case "envfile_global_2.env":
-				return ToReadCloser("COSMOS=STARS"), nil
-			case "envfile_default_profile.env":
-				return ToReadCloser("MAGIC=5"), nil
-			case "envfile_temp_profile.env":
-				return ToReadCloser("WORDS=TIMELESS"), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "testconfig":
+					return ToReadCloser(configData), nil
+				case "envfile_global_1.env":
+					return ToReadCloser("COLOR=BLUE"), nil
+				case "envfile_global_2.env":
+					return ToReadCloser("COSMOS=STARS"), nil
+				case "envfile_default_profile.env":
+					return ToReadCloser("MAGIC=5"), nil
+				case "envfile_temp_profile.env":
+					return ToReadCloser("WORDS=TIMELESS"), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 	app := NewApp(params)
@@ -235,16 +236,18 @@ func TestConfigNotRequiredFile(t *testing.T) {
 	required = true
 `
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "testconfig":
-				return ToReadCloser(configData), nil
-			case "envfile_global_2.env":
-				return ToReadCloser("COSMOS=STARS"), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "testconfig":
+					return ToReadCloser(configData), nil
+				case "envfile_global_2.env":
+					return ToReadCloser("COSMOS=STARS"), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 	app := NewApp(params)
@@ -270,14 +273,16 @@ func TestConfigRequiredFileNotFound(t *testing.T) {
 	required = true
 `
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "testconfig":
-				return ToReadCloser(configData), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "testconfig":
+					return ToReadCloser(configData), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 	app := NewApp(params)
@@ -293,8 +298,10 @@ func fileNeverExists(string) (io.ReadCloser, error) {
 func TestOverwriteDefaultDeny(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+		},
 	}
 	app := NewApp(params)
 
@@ -306,8 +313,10 @@ func TestOverwriteDefaultDeny(t *testing.T) {
 func TestOverwriteExplicitDeny(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+		},
 	}
 	app := NewApp(params)
 
@@ -319,8 +328,10 @@ func TestOverwriteExplicitDeny(t *testing.T) {
 func TestOverwriteExplicitWarn(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+		},
 	}
 	app := NewApp(params)
 
@@ -333,8 +344,10 @@ func TestOverwriteExplicitWarn(t *testing.T) {
 func TestOverwriteExplicitAllow(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+		},
 	}
 	app := NewApp(params)
 
@@ -347,8 +360,10 @@ func TestOverwriteExplicitAllow(t *testing.T) {
 func TestOverwriteInvalidPolicy(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+		},
 	}
 	app := NewApp(params)
 
@@ -361,14 +376,16 @@ func TestOverwriteEvnFile(t *testing.T) {
 	buf := &bytes.Buffer{}
 
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "my.env":
-				return ToReadCloser("COLOR=BLUE"), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "my.env":
+					return ToReadCloser("COLOR=BLUE"), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 	app := NewApp(params)
@@ -386,14 +403,16 @@ func TestOverwriteSetInConfig(t *testing.T) {
 	overwrite = "allow"
 `
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "testconfig":
-				return ToReadCloser(configData), nil
-			default:
-				return nil, os.ErrNotExist
-			}
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc: func(fname string) (io.ReadCloser, error) {
+				switch fname {
+				case "testconfig":
+					return ToReadCloser(configData), nil
+				default:
+					return nil, os.ErrNotExist
+				}
+			},
 		},
 	}
 	app := NewApp(params)
@@ -407,9 +426,11 @@ func TestOverwriteSetInConfig(t *testing.T) {
 func TestCommandPrompt(t *testing.T) {
 	buf := &bytes.Buffer{}
 	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
-		InputFunc:    func(string) string { return "BLUE" },
+		ExtIO: &ExtIOFunc{
+			DryRunOutput: buf,
+			OpenFunc:     fileNeverExists,
+			InputFunc:    func(string) string { return "BLUE" },
+		},
 	}
 	app := NewApp(params)
 
@@ -417,272 +438,4 @@ func TestCommandPrompt(t *testing.T) {
 	require.NoError(t, err)
 	envmap := toEnvVars(buf)
 	assert.Equal(t, "BLUE", envmap["COLOR"])
-}
-
-func getKeychainItemString(item keychain.Item, key string) string {
-	value := reflect.ValueOf(item)
-	vdata := value.FieldByName("attr")
-	val := vdata.MapIndex(reflect.ValueOf(key))
-
-	if val.IsValid() {
-		return fmt.Sprintf("%v", val)
-	} else {
-		return ""
-	}
-}
-
-func TestKeyChainPut(t *testing.T) {
-	buf := &bytes.Buffer{}
-	callAdd, callUpdate, callQuery := false, false, false
-	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
-		KeychainAddItem: func(item keychain.Item) error {
-			callAdd = true
-			assert.Equal(t, "altenv.ns1", getKeychainItemString(item, keychain.ServiceKey))
-			account := getKeychainItemString(item, keychain.AccountKey)
-			switch account {
-			case "COLOR":
-				return keychain.ErrorDuplicateItem // dup
-			case "MAGIC":
-				return nil // success
-			default:
-				require.Failf(t, "Inavlid account key: %s", account)
-			}
-
-			return nil
-		},
-		KeychainUpdateItem: func(query keychain.Item, item keychain.Item) error {
-			// Update only COLOR by keychain.ErrorDuplicateItem
-			callUpdate = true
-			assert.Equal(t, "altenv.ns1", getKeychainItemString(query, keychain.ServiceKey))
-			assert.Equal(t, "COLOR", getKeychainItemString(query, keychain.AccountKey))
-			assert.Equal(t, "altenv.ns1", getKeychainItemString(item, keychain.ServiceKey))
-			assert.Equal(t, "COLOR", getKeychainItemString(item, keychain.AccountKey))
-
-			return nil
-		},
-		KeychainQueryItem: func(query keychain.Item) ([]keychain.QueryResult, error) {
-			callQuery = true
-			return nil, nil
-		},
-	}
-	app := NewApp(params)
-
-	err := app.Run([]string{"altenv",
-		"-r", "update-keychain",
-		"-d", "COLOR=BLUE", "-d", "MAGIC=5",
-		"-w", "ns1",
-	})
-	require.NoError(t, err)
-	//	envmap := toEnvVars(buf)
-	assert.True(t, callAdd)
-	assert.True(t, callUpdate)
-	assert.False(t, callQuery)
-}
-
-func TestKeyChainGet(t *testing.T) {
-	buf := &bytes.Buffer{}
-	callAdd, callUpdate, callQueryAll, callQueryOne := 0, 0, 0, 0
-	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc:     fileNeverExists,
-		KeychainAddItem: func(item keychain.Item) error {
-			callAdd++
-			return nil
-		},
-		KeychainUpdateItem: func(query keychain.Item, item keychain.Item) error {
-			callUpdate++
-			return nil
-		},
-		KeychainQueryItem: func(query keychain.Item) ([]keychain.QueryResult, error) {
-			assert.Equal(t, "altenv.ns1", getKeychainItemString(query, keychain.ServiceKey))
-			account := getKeychainItemString(query, keychain.AccountKey)
-
-			switch account {
-			case "":
-				callQueryAll++
-				return []keychain.QueryResult{
-					{Account: "COLOR"},
-					{Account: "MAGIC"},
-				}, nil
-			case "COLOR":
-				callQueryOne++
-				return []keychain.QueryResult{{Data: []byte("BLUE")}}, nil
-			case "MAGIC":
-				callQueryOne++
-				return []keychain.QueryResult{{Data: []byte("5")}}, nil
-			}
-
-			return nil, fmt.Errorf("Fail")
-		},
-	}
-	app := NewApp(params)
-
-	err := app.Run(newArgs("-k", "ns1"))
-	require.NoError(t, err)
-	//	envmap := toEnvVars(buf)
-	assert.Equal(t, 0, callAdd)
-	assert.Equal(t, 0, callUpdate)
-	assert.Equal(t, 1, callQueryAll)
-	assert.Equal(t, 2, callQueryOne)
-
-	envmap := toEnvVars(buf)
-	assert.Equal(t, "BLUE", envmap["COLOR"])
-	assert.Equal(t, "5", envmap["MAGIC"])
-}
-
-func TestKeyChainGetByConfig(t *testing.T) {
-	buf := &bytes.Buffer{}
-	callAdd, callUpdate, callQueryAll, callQueryOne := 0, 0, 0, 0
-	configData := `
-[global]
-keychain=["ns1"]
-`
-
-	params := &Parameters{
-		DryRunOutput: buf,
-		OpenFunc: func(fname string) (io.ReadCloser, error) {
-			switch fname {
-			case "testconfig":
-				return ToReadCloser(configData), nil
-			default:
-				return nil, os.ErrNotExist
-			}
-		},
-		KeychainAddItem: func(item keychain.Item) error {
-			callAdd++
-			return nil
-		},
-		KeychainUpdateItem: func(query keychain.Item, item keychain.Item) error {
-			callUpdate++
-			return nil
-		},
-		KeychainQueryItem: func(query keychain.Item) ([]keychain.QueryResult, error) {
-			assert.Equal(t, "altenv.ns1", getKeychainItemString(query, keychain.ServiceKey))
-			account := getKeychainItemString(query, keychain.AccountKey)
-
-			switch account {
-			case "":
-				callQueryAll++
-				return []keychain.QueryResult{
-					{Account: "COLOR"},
-					{Account: "MAGIC"},
-				}, nil
-			case "COLOR":
-				callQueryOne++
-				return []keychain.QueryResult{{Data: []byte("BLUE")}}, nil
-			case "MAGIC":
-				callQueryOne++
-				return []keychain.QueryResult{{Data: []byte("5")}}, nil
-			}
-
-			return nil, fmt.Errorf("Fail")
-		},
-	}
-	app := NewApp(params)
-
-	err := app.Run(newArgs("-c", "testconfig"))
-	require.NoError(t, err)
-	//	envmap := toEnvVars(buf)
-	assert.Equal(t, 0, callAdd)
-	assert.Equal(t, 0, callUpdate)
-	assert.Equal(t, 1, callQueryAll)
-	assert.Equal(t, 2, callQueryOne)
-
-	envmap := toEnvVars(buf)
-	assert.Equal(t, "BLUE", envmap["COLOR"])
-	assert.Equal(t, "5", envmap["MAGIC"])
-}
-
-func TestKeyChainServicePrefix(t *testing.T) {
-	buf := &bytes.Buffer{}
-	callAdd, callQuery := 0, 0
-	newParam := func() *Parameters {
-		return &Parameters{
-			DryRunOutput: buf,
-			OpenFunc:     fileNeverExists,
-			KeychainAddItem: func(item keychain.Item) error {
-				callAdd++
-				assert.Equal(t, "clocktower-ns1", getKeychainItemString(item, keychain.ServiceKey))
-				return nil
-			},
-			KeychainUpdateItem: func(query keychain.Item, item keychain.Item) error { return nil },
-			KeychainQueryItem: func(query keychain.Item) ([]keychain.QueryResult, error) {
-				callQuery++
-				assert.Equal(t, "clocktower-ns1", getKeychainItemString(query, keychain.ServiceKey))
-				switch getKeychainItemString(query, keychain.AccountKey) {
-				case "":
-					return []keychain.QueryResult{{Account: "COLOR"}}, nil
-				case "COLOR":
-					return []keychain.QueryResult{{Data: []byte("BLUE")}}, nil
-				default:
-					return nil, fmt.Errorf("Fail")
-				}
-			},
-		}
-	}
-
-	// Put action
-	app1 := NewApp(newParam())
-	err := app1.Run([]string{"altenv",
-		"-r", "update-keychain",
-		"--keychain-service-prefix", "clocktower-",
-		"-d", "COLOR=BLUE",
-		"-w", "ns1",
-	})
-	require.NoError(t, err)
-	assert.Equal(t, 1, callAdd)
-	assert.Equal(t, 0, callQuery)
-
-	// Get action
-	app2 := NewApp(newParam())
-	err = app2.Run(newArgs("-k", "ns1", "--keychain-service-prefix", "clocktower-"))
-	require.NoError(t, err)
-	assert.Equal(t, 1, callAdd)
-	assert.Equal(t, 2, callQuery)
-	envmap := toEnvVars(buf)
-	assert.Equal(t, "BLUE", envmap["COLOR"])
-}
-
-func TestKeyChainServicePrefixByConfig(t *testing.T) {
-	buf := &bytes.Buffer{}
-	callAdd := 0
-	configData := `
-[global]
-keychainServicePrefix="clocktower-"
-`
-	newParam := func() *Parameters {
-		return &Parameters{
-			DryRunOutput: buf,
-			OpenFunc: func(fname string) (io.ReadCloser, error) {
-				switch fname {
-				case "testconfig":
-					return ToReadCloser(configData), nil
-				default:
-					return nil, os.ErrNotExist
-				}
-			},
-			KeychainAddItem: func(item keychain.Item) error {
-				callAdd++
-				assert.Equal(t, "clocktower-ns1", getKeychainItemString(item, keychain.ServiceKey))
-				return nil
-			},
-			KeychainUpdateItem: func(query keychain.Item, item keychain.Item) error { return nil },
-			KeychainQueryItem: func(query keychain.Item) ([]keychain.QueryResult, error) {
-				return nil, nil
-			},
-		}
-	}
-
-	// Put action
-	app := NewApp(newParam())
-	err := app.Run([]string{"altenv",
-		"-c", "testconfig",
-		"-r", "update-keychain",
-		"-d", "COLOR=BLUE",
-		"-w", "ns1",
-	})
-	require.NoError(t, err)
-	assert.Equal(t, 1, callAdd)
 }
