+++ keychain.go
@@ -1,12 +1,6 @@
 package main
 
-import "github.com/keybase/go-keychain"
-
 const keychainServiceNamePrefix = "altenv."
-
-type keychainAddItem func(keychain.Item) error
-type keychainUpdateItem func(keychain.Item, keychain.Item) error
-type keychainQueryItem func(keychain.Item) ([]keychain.QueryResult, error)
 
 type putKeyChainValuesArgs struct {
 	envvars       []*envvar
