+++ keychain_linux.go
@@ -6,10 +6,20 @@ import (
 	"fmt"
 )
 
-func putKeyChainValues(envvars []*envvar, namespace string) error {
+type keychainAddItem func() error
+type keychainUpdateItem func() error
+type keychainQueryItem func() error
+
+func setupKeychainFunc(extIO *ExtIOFunc) {
+	extIO.KeychainAddItem = func() error { return nil }
+	extIO.KeychainUpdateItem = func() error { return nil }
+	extIO.KeychainQueryItem = func() error { return nil }
+}
+
+func putKeyChainValues(putKeyChainValuesArgs) error {
 	return fmt.Errorf("Keychain is not supported in the OS")
 }
 
-func getKeyChainValues(namespace string) ([]*envvar, error) {
+func getKeyChainValues(getKeyChainValuesArgs) ([]*envvar, error) {
 	return nil, fmt.Errorf("Keychain is not supported in the OS")
 }
