+++ parameters.go
@@ -4,8 +4,6 @@ import (
 	"io"
 	"os"
 
-	"github.com/Songmu/prompter"
-	"github.com/keybase/go-keychain"
 	cli "github.com/urfave/cli/v2"
 )
 
@@ -29,21 +27,11 @@ type parameters struct {
 	KeychainServicePrefix string
 
 	// For testing
-	DryRunOutput       io.Writer
-	OpenFunc           fileOpen
-	InputFunc          promptInput
-	KeychainAddItem    keychainAddItem
-	KeychainUpdateItem keychainUpdateItem
-	KeychainQueryItem  keychainQueryItem
+	ExtIO *ExtIOFunc
 }
 
 func newParameters() parameters {
 	return parameters{
-		DryRunOutput:       os.Stdout,
-		OpenFunc:           wrapOSOpen,
-		InputFunc:          prompter.Password,
-		KeychainAddItem:    keychain.AddItem,
-		KeychainUpdateItem: keychain.UpdateItem,
-		KeychainQueryItem:  keychain.QueryItem,
+		ExtIO: NewExtIOFunc(),
 	}
 }
